fix: circular dependency in CLI
